docs: battle domain 주석 정리 및 DTO req/res별 package화
