+++ src/main/java/Hampouch/server/domain/battle/controller/BattleController.java
@@ -1,6 +1,7 @@
 package Hampouch.server.domain.battle.controller;
 
-import Hampouch.server.domain.battle.dto.*;
+import Hampouch.server.domain.battle.dto.request.CreateBattleRequest;
+import Hampouch.server.domain.battle.dto.response.*;
 import Hampouch.server.domain.battle.entity.BattleStatus;
 import Hampouch.server.domain.battle.service.BattleService;
 import Hampouch.server.global.common.response.ApiResponse;
@@ -24,11 +25,7 @@ public class BattleController {
 
     private final BattleService battleService;
 
-    /**
-     * POST /api/battles — 201 + Location 헤더. 성공 메시지를 덮어쓰는 이유: 저장소의 커스텀
-     * 성공 메시지는 전부 쓰기 동작에만 붙는 패턴이라(auth의 회원가입/로그인, expense의 삭제,
-     * 아래 참가) 생성도 여기에 맞춘다. 조회 계열은 기본 문구 유지.
-     */
+    /** POST /api/battles — 201 + Location 헤더*/
     @PostMapping
     public ResponseEntity<ApiResponse<CreateBattleResponse>> create(
             @LoginUserId Long userId,
@@ -47,11 +44,7 @@ public class BattleController {
         return ApiResponse.success(battleService.getMyBattles(userId, status));
     }
 
-    /**
-     * GET /api/battles/{battleId} — 배틀 상세 + 참가자 랭킹. 참가자만 조회 가능하며 아니면
-     * FORBIDDEN_NOT_PARTICIPANT. 경로 순서상 /invitations/{battleCode}보다 뒤에 둬야 하는 제약은
-     * 없음 — Spring MVC가 정적 세그먼트(/invitations)를 가변 세그먼트({battleId})보다 우선 매칭한다.
-     */
+    /** GET /api/battles/{battleId} — 배틀 상세 + 참가자 랭킹. 참가자만 조회 가능*/
     @GetMapping("/{battleId}")
     public ApiResponse<BattleDetailResponse> getBattleDetail(
             @LoginUserId Long userId,
@@ -59,10 +52,7 @@ public class BattleController {
         return ApiResponse.success(battleService.getBattleDetail(userId, battleId));
     }
 
-    /**
-     * GET /api/battles/invitations/{battleCode} — 참가 전 미리보기(로그인 필수).
-     * 참가 가능 여부는 응답 필드가 아니라 에러(BATTLE_FULL 등)로 분기하므로 200이면 곧 참가 가능.
-     */
+    /**GET /api/battles/invitations/{battleCode} — 참가 링크 눌렀을 때 햄배틀 정보 출력 화면.*/
     @GetMapping("/invitations/{battleCode}")
     public ApiResponse<BattleInvitationResponse> getInvitation(
             @LoginUserId Long userId,
@@ -70,9 +60,7 @@ public class BattleController {
         return ApiResponse.success(battleService.getInvitation(userId, battleCode));
     }
 
-    /**
-     * POST /api/battles/invitations/{battleCode} — 참가. Location 헤더와 body 양쪽에 battleId를 싣는다
-     */
+    /**POST /api/battles/invitations/{battleCode} — 참가. Location 헤더와 body 양쪽에 battleId를 싣는다*/
     @PostMapping("/invitations/{battleCode}")
     public ResponseEntity<ApiResponse<JoinBattleResponse>> join(
             @LoginUserId Long userId,
