+++ src/main/java/Hampouch/server/domain/battle/dto/response/BattleSummary.java
@@ -1,4 +1,4 @@
-package Hampouch.server.domain.battle.dto;
+package Hampouch.server.domain.battle.dto.response;
 
 import Hampouch.server.domain.battle.entity.BattleStatus;
 
@@ -7,8 +7,7 @@ import java.util.List;
 
 /**
  * GET /battles 목록 항목 — READY(정원/참가인원)·ONGOING(참가자별 today/total)·TERMINATED(승자 요약)
- * 3가지가 shape 자체가 달라, 공통 필드+널러블 옵션필드 하나 대신 sealed interface로 나눔
- * (status별 필드 누락·오조합을 컴파일 타임에 막음, Java 21).
+ * 3가지가 shape 자체가 달라, 공통 필드+널러블 옵션필드 하나 대신 sealed interface로 분리.
  */
 public sealed interface BattleSummary
         permits BattleSummary.Ready, BattleSummary.Ongoing, BattleSummary.Terminated {
