+++ src/main/java/Hampouch/server/domain/battle/service/BattleBatchScheduler.java
@@ -11,10 +11,8 @@ import java.time.Clock;
 import java.time.LocalDate;
 
 /**
- * 배틀 자동 취소·무효화·종료 스냅샷 배치의 스케줄러(#139). 매일 시작→무효화→종료 순서로 실행한다
- * — 무효화(탈퇴 유저 즉시 무효화 포함)가 종료보다 먼저 돌아야 그날 바로 종료되는 배틀의 랭킹·벌칙
- * 대상 산정에 즉시 반영된다. ChallengeFinalizationScheduler와 동일하게 대상 하나의 실패가 나머지를
- * 막지 않도록 항목별로 예외를 잡아 로그만 남긴다.
+ * 배틀 자동 취소·무효화·종료 스냅샷 배치의 스케줄러. 매일 시작→무효화→종료 순서로 실행
+ * 무효화가 종료보다 먼저 돌아야 그날 종료되는 배틀의 랭킹·벌칙 대상 산정에 즉시 반영된다.
  */
 @Slf4j
 @Component
